JavaReturnExample
In this project, we will show how to use the return statement in methods in Java. The return statement is used to return the result of a method back to where it was called. This example will show how to use return to return the result of a method that squares a number and how to use return with different data types.